stellar/image_resize_requirements.txt
Pillow==11.0.0
psd_tools==1.10.4
Send2Trash==1.8.3
